Move temporary Cargo.toml modification logic to a separate module
