+++ xtask/src/extendrtests/mod.rs
@@ -1,0 +1,1 @@
+pub(crate) mod with_absolute_path;
